requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
python-jose[cryptography]
passlib[argon2]
pytest
pytest-asyncio
httpx
aiosqlite
alembic